requirements.txt
typing
tensorflow
tensorflow_addons
numpy
future
fastapi
ffmpeg
ffmpeg-python
temp
tqdm
Pillow
uvicorn
shutils
PyYAML
absl-py
tensorflow-gpu
pandas
openpyxl
protobuf
python-multipart
sentence_transformers
fifty
imagehash
redis
redisai